backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
exceptiongroup==1.3.0
fastapi==0.115.12
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
